Added user avatar upload and display on users card

+++ SmartEnergy/SmartEnergy.Service/Services/MultimediaService.cs
@@ -3,17 +3,22 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using nClam;
+using SmartEnergy.Contract.CustomExceptions;
 using SmartEnergy.Contract.CustomExceptions.Multimedia;
+using SmartEnergy.Contract.CustomExceptions.User;
 using SmartEnergy.Contract.CustomExceptions.WorkRequest;
 using SmartEnergy.Contract.DTO;
+using SmartEnergy.Contract.Enums;
 using SmartEnergy.Contract.Interfaces;
 using SmartEnergy.Infrastructure;
 using SmartEnergyDomainModels;
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace SmartEnergy.Service.Services
@@ -39,6 +44,9 @@ namespace SmartEnergy.Service.Services
             if (wr == null)
                 throw new WorkRequestNotFound($"Work request with id {workRequestId} does not exist.");
 
+            if (wr.DocumentStatus == DocumentStatus.APPROVED || wr.DocumentStatus == DocumentStatus.CANCELLED)
+                throw new WorkRequestInvalidStateException($"Cannot attach to this work request as it is already {wr.DocumentStatus}");
+
             string filePath = Path.Combine(@$"Attachments/WR{workRequestId}/", formFile.FileName);
             if(_dbContext.MultimediaAttachments.FirstOrDefault(x => x.MultimediaAnchorID == wr.MultimediaAnchorID
                                                                && x.Url == formFile.FileName) != null)
@@ -62,6 +70,36 @@ namespace SmartEnergy.Service.Services
 
         }
 
+        public async Task AttachUserAvatar(IFormFile formFile, int userId)
+        {
+            await ScanAttachmentAsync(formFile);
+
+            User user = _dbContext.Users.Find(userId);
+
+            if (!IsImage(formFile))
+                throw new MultimediaNotImageException($"Uploaded file is not an image!");
+
+            if (user == null)
+                throw new UserNotFoundException($"User with ID {userId} does not exist.");
+
+            string filePath = Path.Combine(@$"Attachments/Users/User{userId}/", formFile.FileName);
+            if(Directory.Exists(@$"Attachments/Users/User{userId}"))
+            {
+                Directory.Delete(@$"Attachments/Users/User{userId}");
+            }
+
+            new FileInfo(filePath).Directory?.Create();
+            using (FileStream stream = new FileStream(filePath, FileMode.Create))
+            {
+                formFile.CopyTo(stream);
+
+            }
+
+            user.ImageURL = formFile.FileName;
+
+            _dbContext.SaveChanges();
+        }
+
         public void DeleteWorkRequestAttachment(int workRequestId, string filename)
         {
             WorkRequest wr = _dbContext.WorkRequests.Find(workRequestId);
@@ -69,11 +107,14 @@ namespace SmartEnergy.Service.Services
             if (wr == null)
                 throw new WorkRequestNotFound($"Work request with id {workRequestId} does not exist.");
 
+            if (wr.DocumentStatus == DocumentStatus.APPROVED || wr.DocumentStatus == DocumentStatus.CANCELLED)
+                throw new WorkRequestInvalidStateException($"Cannot delete attachment from this work request as it is already {wr.DocumentStatus}");
+
             MultimediaAttachment attachment = _dbContext.MultimediaAttachments.FirstOrDefault(x => x.MultimediaAnchorID == wr.MultimediaAnchorID
                                                                                                 && x.Url == filename);
 
             if (attachment == null)
-                throw new MultimediaNotFoundException($"Work request with ID {workRequestId} does not contain file wiht name {filename}");
+                throw new MultimediaNotFoundException($"Work request with ID {workRequestId} does not contain file with name {filename}");
 
             _dbContext.MultimediaAttachments.Remove(attachment);
 
@@ -83,6 +124,19 @@ namespace SmartEnergy.Service.Services
             }
 
             _dbContext.SaveChanges();
+        }
+
+        public FileStream GetUserAvatarStream(int userId, string imageURL)
+        {
+            User user = _dbContext.Users.Find(userId);
+            if (user == null)
+                throw new UserNotFoundException($"User with id {userId} does not exist");
+
+            if (user.ImageURL != imageURL)
+                throw new MultimediaNotFoundException($"User does not have profile picture!");
+
+            FileStream stream = new FileStream(@$"Attachments/Users/User{userId}/{imageURL}", FileMode.Open);
+            return stream;
         }
 
         public List<MultimediaAttachmentDto> GetWorkRequestAttachments(int workRequestId)
@@ -123,6 +177,87 @@ namespace SmartEnergy.Service.Services
             var scanResult = await clam.SendAndScanFileAsync(fileBytes);
            if (scanResult.Result != ClamScanResults.Clean)
                 throw new MultimediaInfectedException($"This attachment is infected with virus!");
+        }
+
+        private bool IsImage(IFormFile postedFile)
+        {
+            const int ImageMinimumBytes = 512;
+            //-------------------------------------------
+            //  Check the image mime types
+            //-------------------------------------------
+            if (postedFile.ContentType.ToLower() != "image/jpg" &&
+                        postedFile.ContentType.ToLower() != "image/jpeg" &&
+                        postedFile.ContentType.ToLower() != "image/pjpeg" &&
+                        postedFile.ContentType.ToLower() != "image/gif" &&
+                        postedFile.ContentType.ToLower() != "image/x-png" &&
+                        postedFile.ContentType.ToLower() != "image/png")
+            {
+                return false;
+            }
+
+            //-------------------------------------------
+            //  Check the image extension
+            //-------------------------------------------
+            if (Path.GetExtension(postedFile.FileName).ToLower() != ".jpg"
+                && Path.GetExtension(postedFile.FileName).ToLower() != ".png"
+                && Path.GetExtension(postedFile.FileName).ToLower() != ".gif"
+                && Path.GetExtension(postedFile.FileName).ToLower() != ".jpeg")
+            {
+                return false;
+            }
+
+            //-------------------------------------------
+            //  Attempt to read the file and check the first bytes
+            //-------------------------------------------
+            try
+            {
+                if (!postedFile.OpenReadStream().CanRead)
+                {
+                    return false;
+                }
+                //------------------------------------------
+                //check whether the image size exceeding the limit or not
+                //------------------------------------------ 
+                if (postedFile.Length < ImageMinimumBytes)
+                {
+                    return false;
+                }
+
+                byte[] buffer = new byte[ImageMinimumBytes];
+                postedFile.OpenReadStream().Read(buffer, 0, ImageMinimumBytes);
+                string content = System.Text.Encoding.UTF8.GetString(buffer);
+                if (Regex.IsMatch(content, @"<script|<html|<head|<title|<body|<pre|<table|<a\s+href|<img|<plaintext|<cross\-domain\-policy",
+                    RegexOptions.IgnoreCase | RegexOptions.CultureInvariant | RegexOptions.Multiline))
+                {
+                    return false;
+                }
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+
+            //-------------------------------------------
+            //  Try to instantiate new Bitmap, if .NET will throw exception
+            //  we can assume that it's not a valid image
+            //-------------------------------------------
+
+            try
+            {
+                using (var bitmap = new Bitmap(postedFile.OpenReadStream()))
+                {
+                }
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+            finally
+            {
+                postedFile.OpenReadStream().Position = 0;
+            }
+
+            return true;
         }
 
     }
